+++ darksite/darksite/settings.py
@@ -46,6 +46,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
 
+    'adminsortable2',
     'corsheaders',
     'graphene_django',