+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Added blog post model.
 * Added endpoint to fetch post list.
 * Added media resource model.
+* Added model to track content of information panels shown on home screen.